ycy/requirements.txt
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bigdl-llm==2.4.0b20230920
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi @ file:///C:/b/abs_36eb5mzhph/croot/certifi_1690232276943/work/certifi
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
dataclasses-json==0.6.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.7.4
fastjsonschema==2.18.0
filelock==3.12.4
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.1
gitdb==4.0.10
GitPython==3.1.36
greenlet==2.0.2
huggingface-hub==0.17.2
idna==3.4
importlib-metadata==6.8.0
InstructorEmbedding==1.0.1
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
langchain==0.0.297
langsmith==0.0.38
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
notebook==7.0.4
notebook_shim==0.2.3
numexpr==2.8.6
numpy==1.26.0
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pycparser==2.21
pydantic==1.10.9
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pypdf==3.16.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
redlines==0.4.2
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
rich-click==1.6.1
rpds-py==0.10.3
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.2
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
stack-data==0.6.2
streamlit==1.26.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.3.0
urllib3==2.0.5
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
yarl==1.9.2
zipp==3.17.0
